requirements.txt
Pillow==7.1.2
django-crispy-forms==1.9.0
Django==3.0.5
whitenoise==5.0.1
psycopg2==2.8.5
gunicorn==20.0.4
dj-database-url==0.5.0
django-heroku==0.3.1